handle packs disappearing between dir list and open
What happens is that we get a list of packfiles from Dir.open but
the files can disappear before the File.open on the .idx or .pack
file due to git-repack running at the same time. When this occurs,
with_idx or with_pack would blow up with a NoMethodError because it
tries to close the file in an ensure block, but the file was never
opened.

This change causes the with_idx and with_pack methods to detect this
situation and not yield to the block. This ought to result in
correct behavior (the pack isn't really there) in most
circumstances.