ruby rev_list passes --verify to native rev_parse in fallback
Otherwise, the git-rev-parse will return whatever is given as an arg
when the ref doesn't exist. e.g.,

  $ git rev-parse some-bad-ref
  some-bad-ref
  fatal: ambiguous argument 'some-bad-ref': unknown revision or path not in the working tree.

The error message is on stderr and git-rev-parse exits with non-zero
but the ref name is still output.

The problem here is that code often calls rev_list like:

    git.rev_list({}, "some-bad-ref")

Then rev_list tries to convert some-bad-ref to a SHA1, gets back the
ref string, but continues on anyway. This eventually results in the
rev_list failing to look up the object because it assumes its a SHA1
when its really a ref string.